feat/authorization initial commit
implement authorization on all features

chore/improving type safety

+++ src/app/transaction/details/[purchaseId]/edit/_action/deletePurchase.action.ts
@@ -5,7 +5,10 @@ import { DrizzleError, eq } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 
 import db from "@/infrastructure/database/db";
-import { purchasedItems, purchases } from "@/lib/schema/schema";
+import { purchases } from "@/lib/schema/schema";
+import { auth } from "@/auth";
+import { getUserInfo } from "@/lib/utils/auth";
+import { user } from "@/lib/schema/user";
 
 export async function deletePurchase(
   formData: FormData
@@ -15,30 +18,64 @@ export async function deletePurchase(
   const schema = z.string();
 
   let invariantError: string | undefined;
+
+  const allowedRole: AvailableUserRole[] = ["admin", "manager"];
+
   try {
-    const purchaseId = schema.parse(purchaseIdRaw);
+    const { userId } = await getUserInfo();
+    const { data: purchaseId } = schema.safeParse(purchaseIdRaw);
+    if (!purchaseId) {
+      invariantError = "Invalid Payload";
+      throw new Error(invariantError);
+    }
 
     await db.transaction(async (tx) => {
+      // Validate user role
+      const [userData] = await tx
+        .select({
+          role: user.role,
+        })
+        .from(user)
+        .where(eq(user.id, userId));
+      if (!allowedRole.includes(userData.role)) {
+        invariantError = "Not Allowed";
+        tx.rollback();
+      }
+
+      // Validate current purchase
       const currentPurchase = await tx
         .select()
         .from(purchases)
         .where(eq(purchases.id, purchaseId));
-
       if (currentPurchase.length == 0) {
-        invariantError = "invalid id";
+        invariantError = "invalid purchase id";
         tx.rollback();
       }
 
+      // Validate authorization on current purchase
+      const { creatorId, ownerId } = currentPurchase[0];
+      if (![creatorId, ownerId].includes(userId)) {
+        invariantError = "Not Allowed";
+        tx.rollback();
+      }
+
+      // Commit action to database
       await tx.delete(purchases).where(eq(purchases.id, purchaseId));
     });
 
     revalidatePath(`/transaction/purchase`);
     return { message: `Purchase transaction deleted` };
   } catch (error) {
+    if (invariantError) {
+      return { error: invariantError };
+    }
+
+    if (error instanceof Error && error.message == "USER_LOGGED_OUT") {
+      return { error: "Login first" };
+    }
+
     if (error instanceof DrizzleError) {
-      return {
-        error: invariantError ? invariantError : error.message,
-      };
+      return { error: error.message };
     }
 
     return { error: "internal error" };
